perf(useFela): optimize passing of theme & props
In many simple use cases, the `useFela` hook will be used without any
props passed to it, and will only make the `theme` available to the
rule callbacks in the returned `css` function. In light of that,
this commit makes two optimizations:

  1. Only merge props into the object that is passed to `css`'s
     callbacks when props are actually passed.
  2. Merge props into the `propsWithTheme` object using mutative
     `Object.assign`, as it is by far the most performant way of doing
     so. See: https://jsperf.com/object-assign-vs-spreading/3

These may seem like micro optimizations, but since useFela is a hot
path that could potentially run thousands of times, I think this is
of value in this case.